Moved Discord Functions to module
